requirements.txt
# Core Streamlit and Data Processing
streamlit>=1.24.0
pandas<2.0.0
numpy<1.24.0

# Database connectivity
sqlalchemy<2.0.0
psycopg2-binary

# Excel/File handling
openpyxl
xlrd
xlsxwriter

# String matching and fuzzy search
fuzzywuzzy[speedup]
python-Levenshtein

# LangChain and AI components
langchain>=0.0.184
langchain-community
langchain-core
langchain-ollama

# Vector store and embeddings
faiss-cpu<1.7.0

# Document processing and conversion
python-docx>=0.8.11
reportlab>=4.0.4
Pillow>=10.0.0  # Required by reportlab
pypdf2>=3.0.0   # PDF manipulation
docx2pdf; platform_system=="Windows"  # Optional: for Windows-specific conversion

